# Cactusref optimize cycle detection