--- a/cactusref/Cargo.toml
+++ b/cactusref/Cargo.toml
@@ -6,9 +6,14 @@ edition = "2018"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
+[[bench]]
+name = "drop"
+harness = false
+
 [dependencies]
 log = "0.4.6"
 
 [dev-dependencies]
+criterion = "0.2.11"
 env_logger = "0.6.1"
 libc = "0.2"